Tiny, stupid, dirty but necessary fix
Sometimes EasyPHP will throw an error claiming that it is not finding
"Simplon\Mai" class, but does it only in that piece of code I've
modified, so I'm forcing the class name.
Hopefully we will discover why on earth it is happening and get back the
code to its original version.